Clean up jaf with this commit
Added new automation tools and scripts to help when building
the project. Also cleaned up files struct and contents to better
reflect a good starting pointing for a project. Added .gitemtpyies
to dires so that I can pull them on to the mac and fill them, will
delete when they have something in them.

Correct grunt file problems with placing libs in the wrong area,
still need to work out the screenshot issue and the a management
script for deploy