Move general updates over from areate jaf
Merged all of the updates and new files from the work
being done on the Areate SPA platform. Added the following
to tools and to help with development and production deployment:

- Vagrantfile
- Fabric and the management files for production/ tested only on vagrant
- New templates for Nginx, database, pip and exports

Corrected a lot of smaill things across several files but there still
may be unseen issues. There are also places where code can be more idiomatic